Implement a very basic feature toggle
We now have a redis backed feature toggle system and
we are using it to test singleflight request to get
config json vs regular request.
The first approach returns singleflight overhead makes
server slower, however we need to first improve the
feature toggle system and test the full system before
jump to a conclusion.
The hypothesis is: Reducing Database roundtrips should help
improve overall performace. But if redis is fast enough queueing
request might just be unnecesary overhead. I'll have to work
on a better test scenario to be sure